Separated app render methods
